Improve the on the fly call graph code a bit
Note that the goal of this commit is to make the testcases pass. That is,
it is perfectly possible that this commit just fixes the symptoms instead
of the root causes. For instance, it is unclear (to me!) why the use of a
LinkedHashMultimap instead of a HashMultimap makes a difference in the
WeightedPAutomaton class. This whole commit is basically trial and error
in combination with reading code and doing "educated" guesses (that's also
one reason why this commit message is pretty concise).

Big FAT warning: the on the fly call graph option is an experimental feature.